+++ app/oryx-app-serving/src/main/java/com/cloudera/oryx/app/serving/als/CosineDistanceSensitiveFunction.java
@@ -27,6 +27,6 @@ public interface CosineDistanceSensitiveFunction extends ToDoubleFunction<float[
   /**
    * @return vector whose cosine distance to other vectors influence the function value
    */
-  double[] getTargetVector();
+  float[] getTargetVector();
 
 }
